Fix elibc compiler issues
- Export oeelibc as part of openenclave-targets for install

- Fix elibc/intstr.c sized to unsized type conversions

- Fix elibc/printf.c compiler warnings
   - Fix size of flag definitions to uint32_t
   - Change _write to return size_t rather than ssize_t to match usage
   - Handle oe_vsnprintf and elibc_strtoul error cases
   - Fix various implicit typecasts

- Mark unused parameters in elibc/pthread.c

- Fix unsafe implicit typecasts in elibc/strtoul.c

- Add explicit cast to int for return values of elibc_toupper/tolower

+++ elibc/intstr.c
@@ -49,7 +49,7 @@ const char* elibc_uint64_to_octstr(elibc_intstr_buf_t* buf, uint64_t x, size_t* 
     } while (x /= 8);
 
     if (size)
-        *size = end - p;
+        *size = (size_t)(end - p);
 
     return p;
 }
@@ -68,7 +68,7 @@ const char* elibc_uint64_to_decstr(elibc_intstr_buf_t* buf, uint64_t x, size_t* 
     } while (x /= 10);
 
     if (size)
-        *size = end - p;
+        *size = (size_t)(end - p);
 
     return p;
 }
@@ -104,7 +104,7 @@ const char* elibc_int64_to_decstr(elibc_intstr_buf_t* buf, int64_t x, size_t* si
         *--p = '-';
 
     if (size)
-        *size = end - p;
+        *size = (size_t)(end - p);
 
     return p;
 }
